http service api, dto, mapper

+++ src/main/java/pl/bodzioch/damian/model/ServiceProvider.java
@@ -7,11 +7,11 @@ import lombok.Getter;
 @Getter
 public enum ServiceProvider {
 
-    BTLA_SKILLS(46039, "BTLA SKILLS", 5272872314L),
-    BOOKING_ANIMAL_SPA(124220, "BOOKING ANIMAL SPA", 7162826739L),
-    OPUS_MAGNUS(137771, "Opus Magnus", 9542469063L);
+    BTLA_SKILLS(46039L, "BTLA SKILLS", 5272872314L),
+    BOOKING_ANIMAL_SPA(124220L, "BOOKING ANIMAL SPA", 7162826739L),
+    OPUS_MAGNUS(137771L, "Opus Magnus", 9542469063L);
 
-    private final int id;
+    private final long id;
     private final String name;
     private final long nip;
 }
